Break import cycles at utils/state.js -> form/ui.js
Move enum value out of ui.js

# Conflicts:
#	lib/actions/ui.js
#	lib/components/app/view-switcher.js
#	lib/components/map/connected-stops-overlay.js
#	lib/components/mobile/main.js
#	lib/components/mobile/options-screen.tsx
