+++ lib/actions/form.js
@@ -7,12 +7,8 @@ import qs from 'qs'
 
 import { queryIsValid } from '../util/state'
 
-import {
-  MobileScreens,
-  routeTo,
-  setMainPanelContent,
-  setMobileScreen
-} from './ui'
+import { MobileScreens } from './ui-constants'
+import { routeTo, setMainPanelContent, setMobileScreen } from './ui'
 import { routingQuery } from './api'
 import { setLocation } from './map'
