+++ lib/actions/ui-constants.js
@@ -1,0 +1,19 @@
+// UI state enums
+
+export const MainPanelContent = {
+  PATTERN_VIEWER: 12,
+  ROUTE_VIEWER: 1,
+  STOP_VIEWER: 2,
+  TRIP_VIEWER: 9
+}
+
+export const MobileScreens = {
+  RESULTS_SUMMARY: 8,
+  SEARCH_FORM: 3,
+  SET_DATETIME: 7,
+  SET_FROM_LOCATION: 4,
+  SET_INITIAL_LOCATION: 2,
+  SET_OPTIONS: 6,
+  SET_TO_LOCATION: 5,
+  WELCOME_SCREEN: 1
+}
