+++ lib/components/app/app-menu.tsx
@@ -26,7 +26,8 @@ import * as userActions from '../../actions/user'
 import { ComponentContext } from '../../util/contexts'
 import { handleLocaleSelection } from '../../util/locale'
 import { isModuleEnabled, Modules } from '../../util/config'
-import { MainPanelContent, setMainPanelContent } from '../../actions/ui'
+import { MainPanelContent } from '../../actions/ui-constants'
+import { setMainPanelContent } from '../../actions/ui'
 import { StyledIconWrapper } from '../util/styledIcon'
 import startOver from '../util/start-over'
