+++ lib/components/map/connected-stops-overlay.js
@@ -4,6 +4,7 @@ import StopsOverlay from '@opentripplanner/stops-overlay'
 import * as apiActions from '../../actions/api'
 import * as mapActions from '../../actions/map'
 import * as uiActions from '../../actions/ui'
+import { MainPanelContent } from '../../actions/ui-constants'
 
 // connect to the redux store
 
@@ -11,14 +12,14 @@ const mapStateToProps = (state) => {
   const { mainPanelContent, viewedRoute } = state.otp.ui
   const { patternId, routeId } = viewedRoute || {}
   const { patterns, v2 } = state.otp.transitIndex.routes?.[routeId] || {}
-  const visible = mainPanelContent !== uiActions.MainPanelContent.STOP_VIEWER
+  const visible = mainPanelContent !== MainPanelContent.STOP_VIEWER
 
   let minZoom = 15
   let stops = []
 
   if (
-    (mainPanelContent === uiActions.MainPanelContent.ROUTE_VIEWER ||
-      mainPanelContent === uiActions.MainPanelContent.PATTERN_VIEWER) &&
+    (mainPanelContent === MainPanelContent.ROUTE_VIEWER ||
+      mainPanelContent === MainPanelContent.PATTERN_VIEWER) &&
     patterns
   ) {
     // Avoid duplicates.
