+++ lib/components/mobile/batch-search-screen.js
@@ -4,7 +4,8 @@ import { FormattedMessage } from 'react-intl'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 
-import { MobileScreens, setMobileScreen } from '../../actions/ui'
+import { MobileScreens } from '../../actions/ui-constants'
+import { setMobileScreen } from '../../actions/ui'
 import { StyledIconWrapper } from '../util/styledIcon'
 import BatchSettings from '../form/batch-settings'
 import DefaultMap from '../map/default-map'
