+++ lib/components/mobile/date-time-screen.js
@@ -1,14 +1,15 @@
+import { connect } from 'react-redux'
+import { FormattedMessage } from 'react-intl'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
-import { FormattedMessage } from 'react-intl'
-import { connect } from 'react-redux'
 
+import { MobileScreens } from '../../actions/ui-constants'
+import { setMobileScreen } from '../../actions/ui'
 import DateTimeModal from '../form/date-time-modal'
 import PlanTripButton from '../form/plan-trip-button'
-import { MobileScreens, setMobileScreen } from '../../actions/ui'
 
-import MobileNavigationBar from './navigation-bar'
 import MobileContainer from './container'
+import MobileNavigationBar from './navigation-bar'
 
 class MobileDateTimeScreen extends Component {
   static propTypes = {
@@ -19,23 +20,24 @@ class MobileDateTimeScreen extends Component {
     this.props.setMobileScreen(MobileScreens.RESULTS_SUMMARY)
   }
 
-  render () {
+  render() {
     return (
       <MobileContainer>
         <MobileNavigationBar
           backScreen={MobileScreens.SEARCH_FORM}
-          headerText={<FormattedMessage id='components.DateTimeScreen.header' />}
+          headerText={
+            <FormattedMessage id="components.DateTimeScreen.header" />
+          }
           showBackButton
         />
 
-        <div className='options-main-content mobile-padding'>
+        <div className="options-main-content mobile-padding">
           <DateTimeModal />
         </div>
 
-        <div className='options-lower-tray mobile-padding'>
+        <div className="options-lower-tray mobile-padding">
           <PlanTripButton onClick={this._planTripClicked} />
         </div>
-
       </MobileContainer>
     )
   }
@@ -44,11 +46,14 @@ class MobileDateTimeScreen extends Component {
 // connect to the redux store
 
 const mapStateToProps = (state, ownProps) => {
-  return { }
+  return {}
 }
 
 const mapDispatchToProps = {
   setMobileScreen
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(MobileDateTimeScreen)
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(MobileDateTimeScreen)
