+++ lib/components/mobile/location-search.js
@@ -4,7 +4,8 @@ import { injectIntl } from 'react-intl'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 
-import { MobileScreens, setMobileScreen } from '../../actions/ui'
+import { MobileScreens } from '../../actions/ui-constants'
+import { setMobileScreen } from '../../actions/ui'
 import LocationField from '../form/connected-location-field'
 
 import MobileContainer from './container'
