+++ lib/components/mobile/main.js
@@ -2,9 +2,10 @@ import { connect } from 'react-redux'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 
-import * as uiActions from '../../actions/ui'
 import { ComponentContext } from '../../util/contexts'
 import { getActiveSearch } from '../../util/state'
+import { MainPanelContent, MobileScreens } from '../../actions/ui-constants'
+import { setMobileScreen } from '../../actions/ui'
 
 import MobileDateTimeScreen from './date-time-screen'
 import MobileLocationSearch from './location-search'
@@ -13,8 +14,6 @@ import MobileRouteViewer from './route-viewer'
 import MobileStopViewer from './stop-viewer'
 import MobileTripViewer from './trip-viewer'
 import MobileWelcomeScreen from './welcome-screen'
-
-const { MainPanelContent, MobileScreens } = uiActions
 
 class MobileMain extends Component {
   static propTypes = {
@@ -134,7 +133,7 @@ const mapStateToProps = (state, ownProps) => {
 }
 
 const mapDispatchToProps = {
-  setMobileScreen: uiActions.setMobileScreen
+  setMobileScreen: setMobileScreen
 }
 
 export default connect(mapStateToProps, mapDispatchToProps)(MobileMain)
