+++ lib/components/mobile/welcome-screen.js
@@ -1,18 +1,20 @@
+import { connect } from 'react-redux'
+import { injectIntl } from 'react-intl'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
-import { injectIntl } from 'react-intl'
-import { connect } from 'react-redux'
 
-import LocationField from '../form/connected-location-field'
-import DefaultMap from '../map/default-map'
-import { MobileScreens, setMobileScreen } from '../../actions/ui'
+import { MobileScreens } from '../../actions/ui-constants'
 import { setLocationToCurrent } from '../../actions/map'
+import { setMobileScreen } from '../../actions/ui'
+import DefaultMap from '../map/default-map'
+import LocationField from '../form/connected-location-field'
 
-import MobileNavigationBar from './navigation-bar'
 import MobileContainer from './container'
+import MobileNavigationBar from './navigation-bar'
 
 class MobileWelcomeScreen extends Component {
   static propTypes = {
+    intl: PropTypes.object,
     setLocationToCurrent: PropTypes.func,
     setMobileScreen: PropTypes.func
   }
@@ -35,21 +37,21 @@ class MobileWelcomeScreen extends Component {
     }
   }
 
-  render () {
+  render() {
     return (
       <MobileContainer>
-        <MobileNavigationBar headerText='Plan Your Trip' />
-        <div className='welcome-location mobile-padding'>
+        <MobileNavigationBar headerText="Plan Your Trip" />
+        <div className="welcome-location mobile-padding">
           <LocationField
             inputPlaceholder={this.props.intl.formatMessage({
               id: 'components.WelcomeScreen.prompt'
             })}
-            locationType='to'
+            locationType="to"
             onTextInputClick={this._toFieldClicked}
             showClearButton={false}
           />
         </div>
-        <div className='welcome-map'>
+        <div className="welcome-map">
           <DefaultMap onSetLocation={this._locationSetFromPopup} />
         </div>
       </MobileContainer>
@@ -68,6 +70,7 @@ const mapDispatchToProps = {
   setMobileScreen
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(
-  injectIntl(MobileWelcomeScreen)
-)
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(injectIntl(MobileWelcomeScreen))
