+++ lib/components/viewers/viewer-container.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 
-import { MainPanelContent } from '../../actions/ui'
+import { MainPanelContent } from '../../actions/ui-constants'
 
 import RouteViewer from './route-viewer'
 import StopViewer from './stop-viewer'
