+++ lib/reducers/create-otp-reducer.js
@@ -12,7 +12,7 @@ import {
 } from '../util/constants'
 import { getDefaultModes } from '../util/itinerary'
 import { getISOLikeTimestamp } from '../util/state'
-import { MainPanelContent, MobileScreens } from '../actions/ui'
+import { MainPanelContent, MobileScreens } from '../actions/ui-constants'
 
 const { getTransitModes, isTransit } = coreUtils.itinerary
 const { filterProfileOptions } = coreUtils.profile
