+++ lib/util/state.js
@@ -11,7 +11,7 @@ import qs from 'qs'
 import React from 'react'
 import styled from 'styled-components'
 
-import { MainPanelContent } from '../actions/ui'
+import { MainPanelContent } from '../actions/ui-constants'
 import { summarizeQuery } from '../components/form/user-settings'
 import FormattedMode from '../components/util/formatted-mode'
 import Strong from '../components/util/strong-text'
